# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "### Fast weblog built in golang and top of echo. supports React SSR and hydrate.\nlabstack/echoを使った高速なブログエンジン。\nReactのサーバーサイドレンダリングとクライアントサイドの引き継ぎを実装。\n（ぶっちゃけ速いかと言われるとSSR有効時はそうでもない）",
        "title": "parakeet REST API",
        "contact": {},
        "license": {
            "name": "WTFPL",
            "url": "https://github.com/mohemohe/parakeet/blob/master/LICENSE"
        },
        "version": "1.0 Draft"
    },
    "basePath": "/api",
    "paths": {
        "/v1/auth": {
            "get": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "アクセストークンが有効か判定します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.AuthResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    }
                }
            },
            "post": {
                "description": "メールアドレスとパスワードを送信してアクセストークンを取得します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "login",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.AuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.AuthResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/entries": {
            "get": {
                "description": "エントリー一覧を取得します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entry"
                ],
                "summary": "list entries",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "ページネーションのページ数",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "1ページごとの件数",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "検索クエリー",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Entries"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "エントリーを作成します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entry"
                ],
                "summary": "create entry",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Entry"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.EntryResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    }
                }
            }
        },
        "/v1/entries/{id}": {
            "get": {
                "description": "エントリーを取得します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entry"
                ],
                "summary": "get entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "エントリーの 'Mongo ObjectID'",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.EntryResponse"
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "エントリーを更新します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entry"
                ],
                "summary": "update entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "エントリーの 'Mongo ObjectID'",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Entry"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.EntryResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "エントリーを削除します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entry"
                ],
                "summary": "delete entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "エントリーの 'Mongo ObjectID'",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
...
        "/v1/settings/aws/s3": {
            "get": {
                "description": "S3接続を取得します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "get S3 setting",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "S3接続を設定します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "set S3 setting",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Setting"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.KV"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    }
                }
            }
        },
        "/v1/settings/cache/cloudflare": {
            "get": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "Cloudflareのキャッシュ削除設定を取得します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "get Cloudflare purge info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.KV"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "Cloudflareのキャッシュ削除設定を更新します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "set Cloudflare purge info",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Cloudflare"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.KV"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    }
                }
            }
        },
        "/v1/settings/cache/mongodb": {
            "get": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "MongoDB クエリーのキャッシュ設定を取得します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "get MongoDB query cache",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.KV"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "MongoDB クエリーのキャッシュ設定を更新します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "set MongoDB query cache",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.KV"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.KV"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    }
                }
            }
        },
        "/v1/settings/cache/page": {
            "get": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "サーバーサイドレンダリングで生成したHTMLのキャッシュ設定を取得します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "get SSR-ed page cache",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.KV"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "サーバーサイドレンダリングで生成したHTMLのキャッシュ設定を更新します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "set SSR-ed page cache",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.KV"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.KV"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    }
                }
            }
        },
        "/v1/settings/notify/mastodon": {
            "get": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "Mastodonの通知設定を取得します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "get mastodon notification",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.KV"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "Mastodonの通知設定を更新します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "set mastodon notification",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NotifyMastodon"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.KV"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    }
                }
            }
        },
        "/v1/settings/notify/misskey": {
            "get": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "Misskeyの通知設定を取得します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "get misskey notification",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.KV"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "Misskeyの通知設定を更新します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "set misskey notification",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NotifyMisskey"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.KV"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    }
                }
            }
        },
        "/v1/settings/render/server": {
            "get": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "サーバーサイドレンダリングの設定を取得します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "get SSR",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.KV"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "サーバーサイドレンダリングの設定を更新します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "set SSR",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ServerSideRendering"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.KV"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    }
                }
            }
        },
        "/v1/settings/search/mongodb": {
            "get": {
                "description": "MongoDBの検索設定を取得します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "get mongodb regex search",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "MongoDBの検索設定を設定します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "set mongodb regex search",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Setting"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.KV"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.EmptyJson"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.AuthRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "controllers.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
...
        "controllers.EntryResponse": {
            "type": "object",
            "properties": {
                "entry": {
                    "$ref": "#/definitions/models.Entry"
                }
            }
        },
...
        "middlewares.EmptyJson": {
            "type": "object"
        },
...
        "models.Entries": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Entry"
                    }
                },
                "info": {
                    "$ref": "#/definitions/bongo.PaginationInfo"
                }
            }
        },
        "models.Entry": {
            "type": "object",
            "properties": {
                "_created": {
                    "type": "string"
                },
                "_id": {
                    "type": "string"
                },
                "_modified": {
                    "type": "string"
                },
                "author": {
                    "type": "string"
                },
                "body": {
                    "type": "string"
                },
                "draft": {
                    "type": "boolean"
                },
                "find_count": {
                    "type": "integer"
                },
                "tag": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}